+++ mitra_models/build.rs
@@ -1,0 +1,4 @@
+// https://github.com/rust-db/refinery/issues/309
+fn main() {
+    println!("cargo:rerun-if-changed=migrations");
+}