+++ mitra_models/migrations/V0072__media_attachment__change_description_type.sql
@@ -1,0 +1,1 @@
+ALTER TABLE media_attachment ALTER COLUMN description TYPE TEXT;