+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Use FEP-ef61 payment links in FEP-ef61 actor representation.
 - Accept object attachments with type `Audio`.
 - Write message to log when profile lookup by mention name is successful.
+- Change media description max length to 2000.
 
 ### Removed